게시판 작성자 완료

+++ security1/src/main/java/com/cos/security1/repository/PagingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagingRepository extends JpaRepository<Board, Integer> {
+public interface PagingRepository extends JpaRepository<Board, Long> {
     Page<Board> findByTitleContaining(String searchKeyword, Pageable pageable);
 }
